fix(pager): clear clippy -D warnings surfaced after unblocking workspace lint
With the earlier config/env_overlay blockers gone, clippy reached
xai-grok-pager and reported 10 latent violations: collapsible nested ifs in
event_loop/connect_wizard/astra_logo (collapsed to let-chains), an inherent
ProtocolBackend::from_str shadowing std::str::FromStr (converted to a proper
trait impl, zero callers), and two redundant f32->f32 casts in the welcome
color helpers.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ crates/codegen/xai-grok-pager/src/views/welcome/menu.rs
@@ -37,7 +37,7 @@ fn boost_color(base: Color, target: Color, amount: f32) -> Color {
     let (Color::Rgb(br, bg, bb), Color::Rgb(tr, tg, tb)) = (base, target) else {
         return base;
     };
-    let a = amount.clamp(0.0, 1.0) as f32;
+    let a = amount.clamp(0.0, 1.0);
     let lerp = |c: u8, t: u8| -> u8 {
         let v = c as f32 + (t as f32 - c as f32) * a;
         v.round().clamp(0.0, 255.0) as u8
